+++ angular-workspace/projects/ni/nimble-angular/table-column/number-text/tests/nimble-table-column-number-text.directive.spec.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, ViewChild } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NimbleTableModule } from '../../../table/nimble-table.module';
 import { TableColumnSortDirection } from '../../nimble-table-column-base.directive';
-import { NimbleTableColumnNumberTextDirective, NumberTextFormat, TableColumnNumberText } from '../nimble-table-column-number-text.directive';
+import { NimbleTableColumnNumberTextDirective, NumberTextAlignment, NumberTextFormat, TableColumnNumberText } from '../nimble-table-column-number-text.directive';
 import { NimbleTableColumnNumberTextModule } from '../nimble-table-column-number-text.module';
 
 describe('NimbleTableColumnNumberText', () => {
@@ -77,6 +77,7 @@ describe('NimbleTableColumnNumberText', () => {
                         grouping-disabled
                         format="${NumberTextFormat.decimal}"
                         decimal-digits="6"
+                        alignment="${NumberTextAlignment.left}"
                     ></nimble-table-column-number-text>
                 </nimble-table>
             `
@@ -161,6 +162,11 @@ describe('NimbleTableColumnNumberText', () => {
             expect(nativeElement.format).toBe(NumberTextFormat.decimal);
         });
 
+        it('will use template string values for alignment', () => {
+            expect(directive.alignment).toBe(NumberTextAlignment.left);
+            expect(nativeElement.alignment).toBe(NumberTextAlignment.left);
+        });
+
         it('will use template string values for decimalDigits', () => {
             expect(directive.decimalDigits).toBe(6);
             expect(nativeElement.decimalDigits).toBe(6);
@@ -186,6 +192,7 @@ describe('NimbleTableColumnNumberText', () => {
                         [grouping-disabled]="groupingDisabled"
                         [format]="format"
                         [decimal-digits]="decimalDigits"
+                        [alignment]="alignment"
                     ></nimble-table-column-number-text>
                 </nimble-table>
             `
@@ -204,8 +211,9 @@ describe('NimbleTableColumnNumberText', () => {
             public sortIndex: number | null = 0;
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
-            public format: NumberTextFormat = NumberTextFormat.roundToInteger;
+            public format: NumberTextFormat = NumberTextFormat.decimal;
             public decimalDigits: number | null = 9;
+            public alignment: NumberTextAlignment = NumberTextAlignment.left;
         }
 
         let fixture: ComponentFixture<TestHostComponent>;
@@ -389,14 +397,25 @@ describe('NimbleTableColumnNumberText', () => {
         });
 
         it('can be configured with property binding for format', () => {
-            expect(directive.format).toBe(NumberTextFormat.roundToInteger);
-            expect(nativeElement.format).toBe(NumberTextFormat.roundToInteger);
-
-            fixture.componentInstance.format = NumberTextFormat.decimal;
-            fixture.detectChanges();
-
             expect(directive.format).toBe(NumberTextFormat.decimal);
             expect(nativeElement.format).toBe(NumberTextFormat.decimal);
+
+            fixture.componentInstance.format = NumberTextFormat.default;
+            fixture.detectChanges();
+
+            expect(directive.format).toBe(NumberTextFormat.default);
+            expect(nativeElement.format).toBe(NumberTextFormat.default);
+        });
+
+        it('can be configured with property binding for alignment', () => {
+            expect(directive.alignment).toBe(NumberTextAlignment.left);
+            expect(nativeElement.alignment).toBe(NumberTextAlignment.left);
+
+            fixture.componentInstance.alignment = NumberTextAlignment.right;
+            fixture.detectChanges();
+
+            expect(directive.alignment).toBe(NumberTextAlignment.right);
+            expect(nativeElement.alignment).toBe(NumberTextAlignment.right);
         });
 
         it('can be configured with property binding for decimalDigits', () => {
@@ -441,6 +460,7 @@ describe('NimbleTableColumnNumberText', () => {
                         [attr.grouping-disabled]="groupingDisabled"
                         [attr.format]="format"
                         [attr.decimal-digits]="decimalDigits"
+                        [attr.alignment]="alignment"
                     ></nimble-table-column-number-text>
                 </nimble-table>
             `
@@ -459,8 +479,9 @@ describe('NimbleTableColumnNumberText', () => {
             public sortIndex: number | null = 0;
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
-            public format: NumberTextFormat = NumberTextFormat.roundToInteger;
+            public format: NumberTextFormat = NumberTextFormat.decimal;
             public decimalDigits: number | null = 9;
+            public alignment: NumberTextAlignment = NumberTextAlignment.left;
         }
 
         let fixture: ComponentFixture<TestHostComponent>;
@@ -644,14 +665,28 @@ describe('NimbleTableColumnNumberText', () => {
         });
 
         it('can be configured with attribute binding for format', () => {
-            expect(directive.format).toBe(NumberTextFormat.roundToInteger);
-            expect(nativeElement.format).toBe(NumberTextFormat.roundToInteger);
-
-            fixture.componentInstance.format = NumberTextFormat.decimal;
-            fixture.detectChanges();
-
             expect(directive.format).toBe(NumberTextFormat.decimal);
             expect(nativeElement.format).toBe(NumberTextFormat.decimal);
+
+            fixture.componentInstance.format = NumberTextFormat.default;
+            fixture.detectChanges();
+
+            // Setting the attribute to `undefined` (the value backing NumberTextFormat.default) removes
+            // the attribute, which results in the web component's property being set to `null`.
+            // See https://github.com/microsoft/fast/issues/6630.
+            expect(directive.format).toBeFalsy();
+            expect(nativeElement.format).toBeFalsy();
+        });
+
+        it('can be configured with attribute binding for alignment', () => {
+            expect(directive.alignment).toBe(NumberTextAlignment.left);
+            expect(nativeElement.alignment).toBe(NumberTextAlignment.left);
+
+            fixture.componentInstance.alignment = NumberTextAlignment.right;
+            fixture.detectChanges();
+
+            expect(directive.alignment).toBe(NumberTextAlignment.right);
+            expect(nativeElement.alignment).toBe(NumberTextAlignment.right);
         });
 
         it('can be configured with attribute binding for decimalDigts', () => {