+++ packages/nimble-components/src/rich-text/editor/testing/rich-text-editor.pageobject.ts
@@ -5,7 +5,9 @@ import type { ToggleButton } from '../../../toggle-button';
 import type { ToolbarButton } from './types';
 import {
     getTagsFromElement,
-    getLeafContentsFromElement
+    getLeafContentsFromElement,
+    getLastChildElement,
+    getLastChildElementAttribute
 } from '../../models/testing/markdown-parser-utils';
 
 /**
@@ -53,6 +55,18 @@ export class RichTextEditorPageObject {
         await waitForUpdatesAsync();
     }
 
+    public async pressShiftEnterKeysInEditor(): Promise<void> {
+        const editor = this.getTiptapEditor();
+        const shiftEnterEvent = new KeyboardEvent('keydown', {
+            key: keyEnter,
+            shiftKey: true,
+            bubbles: true,
+            cancelable: true
+        });
+        editor!.dispatchEvent(shiftEnterEvent);
+        await waitForUpdatesAsync();
+    }
+
     public async pressTabKeyInEditor(): Promise<void> {
         const editor = this.getTiptapEditor();
         const event = new KeyboardEvent('keydown', {
@@ -79,6 +93,12 @@ export class RichTextEditorPageObject {
     public async clickFooterButton(button: ToolbarButton): Promise<void> {
         const toggleButton = this.getFormattingButton(button);
         toggleButton!.click();
+        await waitForUpdatesAsync();
+    }
+
+    public async clickFooterIconSlot(button: ToolbarButton): Promise<void> {
+        const icon = this.getIconSlot(button);
+        icon!.click();
         await waitForUpdatesAsync();
     }
 
@@ -109,13 +129,23 @@ export class RichTextEditorPageObject {
     }
 
     public async setEditorTextContent(value: string): Promise<void> {
-        let lastElement = this.getTiptapEditor()?.lastElementChild;
-
-        while (lastElement?.lastElementChild) {
-            lastElement = lastElement?.lastElementChild;
-        }
-        lastElement!.parentElement!.textContent = value;
+        const lastElement = this.getEditorLastChildElement();
+        const textNode = document.createTextNode(value);
+        lastElement.parentElement!.appendChild(textNode);
         await waitForUpdatesAsync();
+    }
+
+    public async replaceEditorContent(value: string): Promise<void> {
+        const lastElement = this.getEditorLastChildElement();
+        lastElement.parentElement!.textContent = value;
+        await waitForUpdatesAsync();
+    }
+
+    public getEditorLastChildAttribute(attribute: string): string {
+        return getLastChildElementAttribute(
+            attribute,
+            this.getTiptapEditor() as HTMLElement
+        );
     }
 
     public getEditorFirstChildTagName(): string {
@@ -134,6 +164,28 @@ export class RichTextEditorPageObject {
         return getLeafContentsFromElement(
             this.getTiptapEditor() as HTMLElement
         );
+    }
+
+    public getEditorTagNamesWithClosingTags(): string[] {
+        const tagNames: string[] = [];
+        const tiptapEditor = this.getTiptapEditor();
+
+        const processNode = (node: Node): void => {
+            if (node.nodeType === Node.ELEMENT_NODE) {
+                const el = node as Element;
+                tagNames.push(el.tagName);
+
+                el.childNodes.forEach(processNode);
+
+                tagNames.push(`/${el.tagName}`);
+            }
+        };
+
+        if (tiptapEditor) {
+            processNode(tiptapEditor);
+        }
+
+        return tagNames.slice(1, -1);
     }
 
     public getFormattingButtonTextContent(
@@ -216,5 +268,16 @@ export class RichTextEditorPageObject {
             'nimble-toggle-button'
         );
         return buttons[button];
+    }
+
+    private getEditorLastChildElement(): Element {
+        return getLastChildElement(this.getTiptapEditor() as HTMLElement)!;
+    }
+
+    private getIconSlot(
+        button: ToolbarButton
+    ): HTMLSpanElement | null | undefined {
+        const toggleButton = this.getFormattingButton(button);
+        return toggleButton?.shadowRoot?.querySelector('.start');
     }
 }