+++ packages/nimble-components/src/rich-text/editor/tests/rich-text-editor-matrix.stories.ts
@@ -30,7 +30,7 @@ const metadata: Meta = {
     }
 };
 
-const richTextMarkdownString = '1. **Bold*Italics***';
+const richTextMarkdownString = '1. <https://nimble.ni.dev>\n2. **Bold*Italics***';
 
 export default metadata;
 
@@ -63,7 +63,7 @@ const component = (
         ${() => footerHiddenName} ${() => errorStateName} ${() => placeholderName} ${() => disabledName} 
     </p>
     <${richTextEditorTag}
-        style="margin: 5px 0px; width: 500px;"
+        style="margin: 5px 0px; width: 500px; height: 100px;"
         ?disabled="${() => disabled}"
         ?footer-hidden="${() => footerHidden}"
         ?error-visible="${() => isError}"
@@ -194,6 +194,35 @@ longWordContentInMobileWidth.play = (): void => {
         .querySelector('nimble-rich-text-editor')!
         .setMarkdown(
             'ThisIsALongWordWithoutSpaceToTestLongWordInSmallWidthThisIsALongWordWithoutSpaceToTestLongWordInSmallWidth'
+        );
+};
+
+const newLineWithForceLineBreakContent = `
+This is a line 1\\
+This line enters new line using hardbreak <br> tag
+
+
+This line enters new line in paragraph tag
+
+
+1. Point 1
+    * Sub point 1\\
+      Hard break sub point content
+`;
+
+export const newLineWithForceLineBreakInMobileWidth: StoryFn = createStory(mobileWidthComponent);
+newLineWithForceLineBreakInMobileWidth.play = (): void => {
+    document
+        .querySelector('nimble-rich-text-editor')!
+        .setMarkdown(newLineWithForceLineBreakContent);
+};
+
+export const longLinkInMobileWidth: StoryFn = createStory(mobileWidthComponent);
+longLinkInMobileWidth.play = (): void => {
+    document
+        .querySelector('nimble-rich-text-editor')!
+        .setMarkdown(
+            '<https://www.google.com/search?q=what+is+nimble&rlz=1C1CHBF_enIN1007IN1007&oq=what+is+nimble&aqs=chrome..69i57j0i512l9.2837j1j7&sourceid=chrome&ie=UTF-8>'
         );
 };