+++ packages/nimble-components/src/rich-text/models/markdown-parser.ts
@@ -3,14 +3,19 @@ import {
     defaultMarkdownParser,
     MarkdownParser
 } from 'prosemirror-markdown';
-import { DOMSerializer } from 'prosemirror-model';
+import { DOMSerializer, Schema } from 'prosemirror-model';
+import { anchorTag } from '../../anchor';
 
 /**
  * Provides markdown parser for rich text components
  */
 export class RichTextMarkdownParser {
+    private static readonly updatedSchema = this.getSchemaWithLinkConfiguration();
+
     private static readonly markdownParser = this.initializeMarkdownParser();
-    private static readonly domSerializer = DOMSerializer.fromSchema(schema);
+    private static readonly domSerializer = DOMSerializer.fromSchema(
+        this.updatedSchema
+    );
 
     /**
      * This function takes a markdown string, parses it using the ProseMirror MarkdownParser, serializes the parsed content into a
@@ -43,13 +48,58 @@ export class RichTextMarkdownParser {
         const supportedTokenizerRules = zeroTokenizerConfiguration.enable([
             'emphasis',
             'list',
-            'autolink'
+            'escape',
+            'autolink',
+            'newline'
         ]);
 
+        supportedTokenizerRules.validateLink = href => /^https?:\/\//i.test(href);
+
+        /**
+         * In order to display encoded characters, non-ASCII characters, emojis, and other special characters in their original form,
+         * we bypass the default normalization of link text in markdown-it. This is done because we support only "AutoLink" feature in CommonMark flavor.
+         * "normalizeLinkText" method reference in markdown-it: https://github.com/markdown-it/markdown-it/blob/2b6cac25823af011ff3bc7628bc9b06e483c5a08/lib/index.js#L67C1-L86C2
+         *
+         * We can use the default normalization once hyperlink support is added.
+         * See: https://github.com/ni/nimble/issues/1527
+         */
+        supportedTokenizerRules.normalizeLinkText = url => url;
+
         return new MarkdownParser(
-            schema,
+            this.updatedSchema,
             supportedTokenizerRules,
             defaultMarkdownParser.tokens
         );
+    }
+
+    private static getSchemaWithLinkConfiguration(): Schema {
+        return new Schema({
+            nodes: schema.spec.nodes,
+            marks: {
+                link: {
+                    attrs: {
+                        href: {},
+                        rel: { default: 'noopener noreferrer' }
+                    },
+                    // Inclusive can be updated when hyperlink support added
+                    // See: https://github.com/ni/nimble/issues/1527
+                    inclusive: false,
+                    // Excludes can be removed/enabled when hyperlink support added
+                    // See: https://github.com/ni/nimble/issues/1527
+                    excludes: '_',
+                    toDOM(node) {
+                        return [
+                            anchorTag,
+                            {
+                                href: node.attrs.href as Attr,
+                                rel: node.attrs.rel as Attr
+                            }
+                        ];
+                    }
+                },
+                em: schema.spec.marks.get('em')!,
+                strong: schema.spec.marks.get('strong')!
+            }
+        });
     }
 }