+++ packages/nimble-components/src/theme-provider/design-token-names.ts
@@ -25,6 +25,7 @@ export const tokenNames: { readonly [key in TokenName]: string } = {
     fillSelectedRgbPartialColor: 'fill-selected-rgb-partial-color',
     fillHoverSelectedColor: 'fill-hover-selected-color',
     fillHoverColor: 'fill-hover-color',
+    fillHoverRgbPartialColor: 'fill-hover-rgb-partial-color',
     fillDownColor: 'fill-down-color',
     borderColor: 'border-color',
     borderRgbPartialColor: 'border-rgb-partial-color',